Squashed commit of the following:
commit 6c725e20bcb2f2fb11e1b32273851f8b17aa29b9
Author: emanuele <me.emanuele.xyz@gmail.com>
Date:   Sat Sep 2 11:35:48 2023 +0200

    don't benckmark by default

commit aa20d885c89a230100638174f892e2f66bfee1ba
Author: emanuele <me.emanuele.xyz@gmail.com>
Date:   Fri Sep 1 14:20:55 2023 +0200

    benchmark build histogram

commit a39fd86b50eb44922f5071342e0d97d3a12385c1
Author: emanuele <me.emanuele.xyz@gmail.com>
Date:   Fri Sep 1 12:36:32 2023 +0200

    do not print arrays

commit 843d69a82e3a08f55c7b4e2cbe06a87ad5600d42
Author: emanuele <me.emanuele.xyz@gmail.com>
Date:   Fri Sep 1 12:36:13 2023 +0200

    remove useless comments

commit 4680e844949da38dfd7e37ef44ae46fa01012cd2
Author: emanuele <me.emanuele.xyz@gmail.com>
Date:   Fri Sep 1 12:35:57 2023 +0200

    remove useless comments

commit bc5947ee7659c3cd5efb8c027390d78833231f29
Author: emanuele <me.emanuele.xyz@gmail.com>
Date:   Fri Sep 1 12:35:31 2023 +0200

    correctly zero initialize shared memory

commit e31284150d43882ae36e70e3cf4f78681ad02764
Author: emanuele <me.emanuele.xyz@gmail.com>
Date:   Fri Sep 1 12:18:19 2023 +0200

    fix write histogram to global memory bug

commit 31f8376461bfe5cdc7d31bcc46d0f6e81118ce63
Author: emanuele <me.emanuele.xyz@gmail.com>
Date:   Thu Aug 31 14:02:11 2023 +0200

    start debugging BuildHistogramsKernel

commit 16a1387b87d84ba4a5af881685a44e42110c277a
Author: emanuele <me.emanuele.xyz@gmail.com>
Date:   Thu Aug 31 13:43:18 2023 +0200

    implement cpu build histograms

commit d599142b0e15acd5762856a885e8bc7315a45e19
Author: emanuele <me.emanuele.xyz@gmail.com>
Date:   Thu Aug 31 13:19:53 2023 +0200

    write boilerplate code

commit 04d8caec438297bc29be603d37dc5c3eee8f0df7
Author: emanuele <me.emanuele.xyz@gmail.com>
Date:   Thu Aug 31 12:57:25 2023 +0200

    write code blockout

commit 952a378e06ec6993d0e7c23f9cff82bd2e8d962b
Author: emanuele <me.emanuele.xyz@gmail.com>
Date:   Thu Aug 31 12:55:02 2023 +0200

    start working on first phase: building the local counter table

commit 5b46590bc0fd8bb1c8faf255f05ff7d903931f6b
Author: emanuele <me.emanuele.xyz@gmail.com>
Date:   Wed Aug 30 13:00:06 2023 +0200

    block out implementations with some details

commit 27b19ea3e4efed15098d7b9e7678f9283ee5aa21
Author: emanuele <me.emanuele.xyz@gmail.com>
Date:   Wed Aug 30 11:57:06 2023 +0200

    set up kernel configuration

commit bc6003e04aa1f16e1f780e83184e5ad1d1fc4777
Author: emanuele <me.emanuele.xyz@gmail.com>
Date:   Wed Aug 30 11:51:39 2023 +0200

    write boilerplate code

commit f80a8e895cb68cdbca0f4997a5f00858b3e72e49
Author: emanuele <me.emanuele.xyz@gmail.com>
Date:   Wed Aug 30 11:15:51 2023 +0200

    blockout code

+++ LSDRadixSort/Utils.cpp
@@ -36,9 +36,9 @@ float GetElapsedMS(int64_t start, int64_t end)
 	return dt * 1000.0f;
 }
 
-void CheckArrays(uint32_t* a, uint32_t* b, int count)
+void CheckArrays(uint32_t* a, uint32_t* b, size_t count)
 {
-	for (int i = 0; i < count; i++)
+	for (size_t i = 0; i < count; i++)
 	{
 		MYASSERT(a[i] == b[i]);
 	}
